- Remove jmp_buf hacking
Set the stack pointer directly (in an architecture-specific way), then
the desired stack pointer is saved to the `jmp_buf` properly.
After this we reset the stack pointer to the original value (if in
calling context) and return normally, while the coroutine context keeps
the updated stack pointer.

This should make things a lot more portable (only dependent on CPU
architecutre but not on OS/libc implementation), but still assumes the
stack grows downwards and the platform has such a thing as a stack
pointer and switching it is enough to switch the execution stack.

Special thanks to @bobsayshilol for a lot of help getting the inline
assembly working, and for helping get research started on porting to
more platforms!